feat(encounter): the last one down ends the fight (#1078)
A bubble left with nobody standing on one side of it dissolves itself, with
cause ByDefeat — the second case of a sealed set this module now owns, mirroring
session's and DCSource's discipline. Ruled fork (c) on #959: the BUBBLE ends,
the encounter does not, and dissolving removes nobody.

Decided inside noticeDown, from the same pull, before the splice: if every
member is down, splicing first drains the bubble and dropBubbleIfIdle deletes
the husk, leaving nothing to end honestly. The news is now written in one pass
and acted on in a second, so an ending never lands before a body that explains
it.

No new event kind and no new persisted state — the existing bubble-dissolved
beat carries the cause, and the pull stays the truth.
