+++ rulebooks/dnd5e/encounter/clocks.go
@@ -527,77 +527,3 @@ func (e *Encounter) EndTurn(in *EndTurnInput) (*EndTurnOutput, error) {
 		Seq:          seq,
 	}, nil
 }
-
-// DissolveInput names a member of the fight being ended. The bubble is
-// reached through a member — never addressed by name — which R6 makes a
-// total function.
-type DissolveInput struct {
-	// Member is anyone in the fight to dissolve.
-	Member MemberID
-}
-
-// DissolveOutput reports who the fight held and where the story recorded
-// its end.
-type DissolveOutput struct {
-	// Members is everyone the fight held, in bubble order, all re-homed to
-	// the world clock at budget zero.
-	Members []MemberID
-
-	// Cause is why the fight ended.
-	Cause DissolveCause
-
-	// Seq is the story sequence of the dissolution beat.
-	Seq uint64
-}
-
-// Dissolve ends the fight holding the named member: the bubble empties and
-// every member re-homes to the world clock, at budget zero — time spent
-// fighting was spent, not banked.
-//
-// Errors: ErrNilInput, ErrClosed, ErrNoMember, ErrNotMember, ErrNoBubble
-// (the member is not in a fight). A failure while re-homing leaves members
-// between clocks — the window ClockOf's on-no-clock check nets — and the
-// caller's obligation is doc.go's: drop the encounter unsaved.
-func (e *Encounter) Dissolve(in *DissolveInput) (*DissolveOutput, error) {
-	if in == nil {
-		return nil, fmt.Errorf("dissolve: %w", ErrNilInput)
-	}
-	if e.outcome != nil {
-		return nil, fmt.Errorf("dissolve: %w", ErrClosed)
-	}
-	if in.Member == "" {
-		return nil, fmt.Errorf("dissolve: %w", ErrNoMember)
-	}
-	if _, ok := e.members[in.Member]; !ok {
-		return nil, fmt.Errorf("dissolve %q: %w", in.Member, ErrNotMember)
-	}
-	bubble, err := e.bubbleFor(in.Member)
-	if err != nil {
-		return nil, fmt.Errorf("dissolve %q: %w", in.Member, err)
-	}
-	if bubble == nil {
-		return nil, fmt.Errorf("dissolve %q: %w", in.Member, ErrNoBubble)
-	}
-
-	out, err := bubble.Dissolve(&clock.DissolveInput{})
-	if err != nil {
-		return nil, fmt.Errorf("dissolve: %w", err)
-	}
-	for _, id := range out.Members {
-		if _, jerr := e.clock.Join(&clock.JoinInput{ID: id}); jerr != nil {
-			return nil, fmt.Errorf("dissolve member %q world clock: %w", id, jerr)
-		}
-	}
-	if derr := e.dropBubbleIfIdle(bubble); derr != nil {
-		return nil, fmt.Errorf("dissolve prune bubble: %w", derr)
-	}
-
-	seq, err := e.appendClockBeat(map[string]interface{}{
-		"beat":    "bubble-dissolved",
-		"members": out.Members,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("dissolve append beat: %w", err)
-	}
-	return &DissolveOutput{Members: out.Members, Seq: seq}, nil
-}
